% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{batch_ukbb_chrpos2rsID}
\alias{batch_ukbb_chrpos2rsID}
\title{调用ukbb数据库的参考数据variants.tsv.gz对ukbb中的GWAS数据批量化进行SNP位点注释}
\usage{
batch_ukbb_chrpos2rsID(ukbb_gwas_dir = NULL, save_dir = NULL)
}
\arguments{
\item{ukbb_gwas_dir}{ukbb格式的GWAS数据所在文件夹}

\item{save_dir}{设置输出数据的文件夹}
}
\value{
返回一个或多个SNP位点注释后的ukbb的GWAS数据
}
\description{
调用ukbb数据库的参考数据variants.tsv.gz对ukbb中的GWAS数据批量化进行SNP位点注释
}
