% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{clump_data_local}
\alias{clump_data_local}
\title{调用ieugwasr包，在本地电脑进行去连锁不平衡，此功能可用于工具变量较多，在线进行clump时总是报错的情况}
\usage{
clump_data_local(
  dat = exposure_dat,
  clump_kb = 10000,
  clump_r2 = 0.001,
  clump_p1 = 1,
  bfile = "D:/生物信息学习/生信自研课程/孟德尔随机化/data/1kg.v3/EUR"
)
}
\arguments{
\item{dat}{TwosampleMR包读入暴露数据的后的数据框}

\item{clump_kb}{暴露中工具变量去连锁不平衡时的步长也叫窗口（windows）的阈值，默认：kb = 10000}

\item{clump_r2}{暴露中工具变量去连锁不平衡时的r2阈值，默认：r2 = 0.001}

\item{clump_p1}{暴露中工具变量去连锁不平衡时的显著性阈值, 默认：clump_p1 = 1}

\item{bfile}{1000基因组数据中作为参考的plink格式人种数据，下载地址：http://fileserve.mrcieu.ac.uk/ld/1kg.v3.tgz，默认：欧洲人群}
}
\value{
返回TwosampleMR包中exposure格式的数据框
}
\description{
调用ieugwasr包，在本地电脑进行去连锁不平衡，此功能可用于工具变量较多，在线进行clump时总是报错的情况
}
