% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{find_anyexposur_outcome}
\alias{find_anyexposur_outcome}
\title{#查询多个暴露对一个结局的阳性结果}
\usage{
find_anyexposur_outcome(
  exposure = exposure,
  outcome = outcome,
  write = F,
  p1 = 5e-08,
  clump = TRUE,
  p2 = 5e-08,
  r2 = 0.001,
  kb = 10000,
  LD = 0.8
)
}
\value{
data frame
}
\description{
#查询多个暴露对一个结局的阳性结果
}
