% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{find_multiexposure_multioutcome_epigraphdb}
\alias{find_multiexposure_multioutcome_epigraphdb}
\title{调用epigraphdb数据库的数据,对多个暴露和/或多个结局进行快速筛选}
\usage{
find_multiexposure_multioutcome_epigraphdb(
  exposure = NULL,
  outcome = NULL,
  pval_threshold = 1e-05,
  write = T,
  save_path = "find_multiexposure_multioutcome_epigraphdb.csv"
)
}
\arguments{
\item{exposure}{暴露的id号或名称，可以是多个，例如：c("ieu-a-2","ieu-a-7")，c("Body mass index","Coronary heart disease")，如果exposure = NUL，则纳入全部的数据}

\item{outcome}{结局的id号或名称，可以是多个，例如：c("ieu-a-2","ieu-a-7")，c("Body mass index","Coronary heart disease")，如果outcome = NUL，则纳入全部的数据}

\item{pval_threshold}{暴露和结局的孟德尔随机化阈值，默认：pval_threshold = 1e-05}

\item{write}{逻辑值，是否将输出结果保存到文件中，默认：write = T 导出结果文件}

\item{save_path}{筛查结果的保存路径和文件名称，默认：save_path = "find_multiexposure_multioutcome_epigraphdb.csv"}
}
\value{
epigraphdb数据库中快速筛查多个暴露和/或多个结局的筛选结果
}
\description{
调用epigraphdb数据库的数据,对多个暴露和/或多个结局进行快速筛选
}
