% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{find_multiexposure_outcome}
\alias{find_multiexposure_outcome}
\title{调用TwoSampleMR包来查询多个暴露对一个结局的阳性结果}
\usage{
find_multiexposure_outcome(
  exposure = exposure,
  outcome = outcome,
  write = T,
  p1 = 5e-08,
  clump = TRUE,
  p2 = 5e-08,
  r2 = 0.001,
  kb = 10000,
  LD = 0.8
)
}
\arguments{
\item{exposure}{暴露的id号，可以是多个，例如：c("ieu-a-2","ieu-a-7")}

\item{outcome}{结局的id号，有且只能是一个}

\item{write}{逻辑值，将初筛结果导出到MR多对一初筛.csv文件中，默认：write=T 导出该文件}

\item{p1}{筛选工具变量的显著性阈值，默认：p1 = 5e-08}

\item{clump}{逻辑值，决定暴露中的工具变量是否进行去连锁不平衡，默认：clump = TRUE}

\item{p2}{TwoSampleMR官网给出的解释是：Secondary clumping threshold，但经我查阅源代码发现这个参数压根没用到(⑅˃◡˂⑅)}

\item{r2}{暴露中工具变量去连锁不平衡时的r2阈值，默认：r2= 0.001}

\item{kb}{暴露中工具变量去连锁不平衡时的步长也叫窗口（windows）的阈值，默认：kb = 10000}

\item{LD}{结局中工具变量的连锁不平衡阈值，即：若暴露中不存在某个工具变量，则寻找该工具变量的代理SNP，LD就是筛选阈值，只有大于该阈值才可作为代理SNP(⑅˃◡˂⑅)}
}
\value{
多个暴露对一个结局的数据框
}
\description{
调用TwoSampleMR包来查询多个暴露对一个结局的阳性结果
}
