% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{find_snp_add_eaf}
\alias{find_snp_add_eaf}
\title{调用ieugwasr包对工具变量中缺失的效应等位基因频率（eaf）进行填充，数据来源于1000基因组数据}
\usage{
find_snp_add_eaf(exposure = NULL, outcome = NULL, pop = "EUR")
}
\arguments{
\item{exposure}{TwosampleMR包读入暴露的数据框}

\item{outcome}{TwosampleMR包读入结局的数据框}

\item{pop}{参考1000基因组数据中的人种}
}
\value{
返回TwosampleMR包中exposure或outcome格式的数据框；注意：有的SNP可能会出现eaf填充失败，这是因为1000基因组数据中缺乏该SNP的eaf信息(⑅˃◡˂⑅)
}
\description{
调用ieugwasr包对工具变量中缺失的效应等位基因频率（eaf）进行填充，数据来源于1000基因组数据
}
