% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendly2MRhelp.R
\name{friendly2MR_update_plans}
\alias{friendly2MR_update_plans}
\title{friendly2MR包的后续更新计划}
\usage{
friendly2MR_update_plans()
}
\value{
列出后续的friendly2MR更新计划
}
\description{
friendly2MR包的后续更新计划
}
\examples{
   friendly2MR_update_plans()
}
