% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{gwasvcf2TwosampleMR_local}
\alias{gwasvcf2TwosampleMR_local}
\title{将从IEU官网下载到本地的gwasvcf文件转换为TwosampleMR的暴露或结局格式}
\usage{
gwasvcf2TwosampleMR_local(
  dat = dat,
  type = "exposure",
  p1 = 5e-08,
  clump = T,
  clump_kb = 10000,
  clump_r2 = 0.001,
  pop = "EUR",
  write = T,
  save_path = "TwosampleMR_formate.txt"
)
}
\arguments{
\item{dat}{从IEU官网下载到本地的gwasvcf文件，例如：IEU-a-2.vcf.gz}

\item{type}{指定该数据作为暴露还是结局，例如：exposure，outcomes，默认：type = "exposure"}

\item{p1}{筛选工具变量的显著性阈值，默认：p1 = 5e-08}

\item{clump}{逻辑值，决定暴露中的工具变量是否进行去连锁不平衡，默认：clump = TRUE}

\item{clump_kb}{暴露中工具变量去连锁不平衡时的步长也叫窗口（windows）的阈值，默认：kb = 10000}

\item{clump_r2}{暴露中工具变量去连锁不平衡时的r2阈值，默认：clump_r2 = 0.001}

\item{pop}{去连锁不平衡时用到的1000人基因组人种，默认：pop = "EUR"}

\item{write}{逻辑值，是否将输出结果保存到文件中，默认：write = T 导出结果文件}

\item{save_path}{筛查结果的保存路径和文件名称，默认：save_path = "TwosampleMR_formate.txt"}
}
\value{
将从IEU官网下载到本地的gwasvcf文件转换为TwosampleMR的暴露或结局格式
}
\description{
将从IEU官网下载到本地的gwasvcf文件转换为TwosampleMR的暴露或结局格式
}
