% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendly2MRhelp.R
\name{install_friendly2MR_dependence}
\alias{install_friendly2MR_dependence}
\title{安装friendly2MR的相关依赖}
\usage{
install_friendly2MR_dependence()
}
\value{
安装friendly2MR的相关依赖包
}
\description{
安装friendly2MR的相关依赖
}
\examples{
   install_friendly2MR_dependence()
}
