% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friendly2MRhelp.R
\name{show_demo_gwas}
\alias{show_demo_gwas}
\title{friendly2MR包中作为输入文件的GWAS数据格式和表头}
\usage{
show_demo_gwas()
}
\value{
输入文件的GWAS数据格式和表头
}
\description{
friendly2MR包中作为输入文件的GWAS数据格式和表头
}
\examples{
   show_demo_gwas()
}
