% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{deletion_confounding_snp}
\alias{deletion_confounding_snp}
\title{定义函数：查找工具变量相关的混杂因素}
\usage{
deletion_confounding_snp(
  confound = NULL,
  exposure_dat = NULL,
  query_gene = NULL,
  query_region = NULL,
  catalogue = "GWAS",
  pvalue = 1e-05,
  proxies = "None",
  r2 = 0.8,
  build = 37,
  write = TRUE,
  filepath = "MR_ivs_confounding.csv"
)
}
\value{
data frame
}
\description{
定义函数：查找工具变量相关的混杂因素
}
